chore: punch app context selection
