+++ libs/punchapp/src/lib/config/index.ts
@@ -1,1 +1,2 @@
 export { WorkspaceWrapper } from './workspaceConfig';
+export { configure } from './frameworkConfig';
